sell価格の小数点が切り捨てられる問題を修正

+++ src/main/java/net/okocraft/box/util/GeneralConfig.java
@@ -103,7 +103,7 @@ public class GeneralConfig {
     @Getter
     private List<String> allItems;
     @Getter
-    private Map<String, Integer> sellPrice;
+    private Map<String, Double> sellPrice;
 
     // CategorySelectionGui Name
     @Getter
@@ -210,7 +210,7 @@ public class GeneralConfig {
         val priceConfig = new CustomConfig(plugin, "sellprice.yml");
         priceConfig.saveDefaultConfig();
         sellPrice = allItems.stream().collect(Collectors.toMap(itemName -> itemName,
-                itemName -> priceConfig.getConfig().getInt(itemName), (i1, i2) -> i1, LinkedHashMap::new));
+                itemName -> priceConfig.getConfig().getDouble(itemName), (i1, i2) -> i1, LinkedHashMap::new));
     }
 
     /**
